Simplify hackbot agent structure into self-contained agents/<name>/ folders
