+++ agents/bug-fix/agent/firefox_tools/bootstrap_firefox.py
