+++ agents/bug-fix/agent/rules/unsupported-config.md
