+++ agents/bug-fix/pyproject.toml
@@ -4,7 +4,6 @@ version = "0.1.0"
 description = "Cloud Run Job image that runs the bug-fix agent for hackbot-api"
 requires-python = ">=3.12"
 dependencies = [
-    "bugbug",
     "hackbot-runtime[claude-sdk]",
     "bugsy",
     "grizzly-framework",
@@ -16,5 +15,4 @@ dependencies = [
 ]
 
 [tool.uv.sources]
-bugbug = { workspace = true }
 hackbot-runtime = { workspace = true }
