+++ libs/hackbot-runtime/hackbot_runtime/mcp/__init__.py
@@ -1,0 +1,1 @@
+"""In-process MCP servers for hackbot agents (claude-sdk extra)."""
